# prowl/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Prowl LLM Orchestration API",
    "description": "REST API for LLM-driven orchestration of the Prowl security crawler. Enables programmatic crawl control, discovery queries, mid-crawl adjustments, intervention handling, and attack surface reporting.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/v1/crawl/start": {
      "post": {
        "tags": [
          "v1",
          "crawl"
        ],
        "summary": "Start a new crawl",
        "description": "Launch a new crawl session with full configuration. The crawl runs as a background task. Use GET /crawl/status to monitor progress.",
        "operationId": "start_crawl_api_v1_crawl_start_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlStartRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/phases/{name}": {
      "get": {
        "tags": [
          "v1",
          "crawl"
        ],
        "summary": "Get phase detail",
        "description": "Returns detail for a specific pipeline phase by name.",
        "operationId": "get_phase_api_v1_phases__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PhaseStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/modules/{name}": {
      "get": {
        "tags": [
          "v1",
          "crawl"
        ],
        "summary": "Get module detail",
        "description": "Returns detail for a specific module by name.",
        "operationId": "get_module_api_v1_modules__name__get",
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModuleStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CrawlStartRequest": {
        "properties": {
          "target_url": {
            "type": "string",
            "title": "Target Url",
            "description": "Target URL to crawl"
          },
          "scope_patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Scope Patterns",
            "description": "Regex patterns to include in scope"
          },
          "exclude_patterns": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Exclude Patterns",
            "description": "Regex patterns to exclude from scope"
          },
          "max_depth": {
            "type": "integer",
            "title": "Max Depth",
            "description": "Maximum crawl depth",
            "default": 10
          },
          "max_requests": {
            "type": "integer",
            "title": "Max Requests",
            "description": "Maximum total requests",
            "default": 10000
          },
          "concurrency": {
            "type": "integer",
            "title": "Concurrency",
            "description": "Concurrent request workers",
            "default": 10
          },
          "request_delay": {
            "type": "number",
            "title": "Request Delay",
            "description": "Base delay between requests (seconds)",
            "default": 0.0
          },
          "request_timeout": {
            "type": "number",
            "title": "Request Timeout",
            "description": "HTTP request timeout (seconds)",
            "default": 30.0
          },
          "user_agent": {
            "type": "string",
            "title": "User Agent",
            "description": "User-Agent header value",
            "default": "Prowl/0.1"
          },
          "follow_redirects": {
            "type": "boolean",
            "title": "Follow Redirects",
            "description": "Follow HTTP redirects",
            "default": true
          },
          "backend": {
            "type": "string",
            "enum": [
              "http",
              "browser",
              "hybrid"
            ],
            "title": "Backend",
            "description": "Request backend type",
            "default": "hybrid"
          },
          "headless": {
            "type": "boolean",
            "title": "Headless",
            "description": "Headless browser mode",
            "default": true
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modules",
            "description": "Modules to run (empty=all). Options: s1_spider, s2_bruteforce, s3_params, s4_js, s5_api, s6_passive, s7_auth, s8_states, s9_infra, s10_tech, s11_input, s12_auth, s13_report"
          },
          "output_dir": {
            "type": "string",
            "title": "Output Dir",
            "description": "Output directory path",
            "default": "./prowl-output"
          },
          "output_formats": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Output Formats",
            "description": "Output formats: json, markdown, html, burp"
          },
          "llm_model": {
            "type": "string",
            "title": "Llm Model",
            "description": "LLM model name (e.g. gpt-4o-mini)",
            "default": ""
          },
          "llm_api_key": {
            "type": "string",
            "title": "Llm Api Key",
            "description": "LLM API key",
            "default": ""
          },
          "coverage_guided": {
            "type": "boolean",
            "title": "Coverage Guided",
            "description": "Enable coverage-guided exploration",
            "default": true
          },
          "seed_scheduling": {
            "type": "boolean",
            "title": "Seed Scheduling",
            "description": "Enable Thompson Sampling seed scheduling",
            "default": true
          },
          "saturation_threshold": {
            "type": "number",
            "title": "Saturation Threshold",
            "description": "Coverage saturation threshold",
            "default": 0.02
          },
          "wordlist_dirs": {
            "type": "string",
            "title": "Wordlist Dirs",
            "description": "Path to directory wordlist file",
            "default": ""
          },
          "wordlist_params": {
            "type": "string",
            "title": "Wordlist Params",
            "description": "Path to parameter wordlist file",
            "default": ""
          },
          "auth_roles": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Auth Roles",
            "description": "Auth role configurations"
          },
          "dashboard": {
            "type": "boolean",
            "title": "Dashboard",
            "description": "Enable web dashboard (already running if using API)",
            "default": false
          },
          "dashboard_port": {
            "type": "integer",
            "title": "Dashboard Port",
            "description": "Dashboard port",
            "default": 8484
          }
        },
        "type": "object",
        "required": [
          "target_url"
        ],
        "title": "CrawlStartRequest",
        "description": "Full crawl configuration for programmatic launch.",
        "examples": [
          {
            "concurrency": 15,
            "max_depth": 5,
            "modules": [
              "s6_passive",
              "s1_spider",
              "s4_js"
            ],
            "target_url": "https://example.com"
          }
        ]
      },
      "CrawlStatusResponse": {
        "properties": {
          "state": {
            "type": "string",
            "title": "State",
            "description": "Engine state: idle, running, paused, stopping, stopped"
          },
          "target": {
            "type": "string",
            "title": "Target",
            "description": "Target URL"
          },
          "elapsed": {
            "type": "number",
            "title": "Elapsed",
            "description": "Elapsed time in seconds"
          },
          "requests_completed": {
            "type": "integer",
            "title": "Requests Completed"
          },
          "requests_failed": {
            "type": "integer",
            "title": "Requests Failed"
          },
          "requests_queued": {
            "type": "integer",
            "title": "Requests Queued"
          },
          "requests_dropped": {
            "type": "integer",
            "title": "Requests Dropped"
          },
          "queue_size": {
            "type": "integer",
            "title": "Queue Size"
          },
          "endpoints_found": {
            "type": "integer",
            "title": "Endpoints Found"
          },
          "unique_urls": {
            "type": "integer",
            "title": "Unique Urls"
          },
          "transactions_stored": {
            "type": "integer",
            "title": "Transactions Stored"
          },
          "phase_name": {
            "type": "string",
            "title": "Phase Name",
            "description": "Current phase name"
          },
          "current_phase": {
            "type": "integer",
            "title": "Current Phase",
            "description": "Current phase index"
          },
          "coverage": {
            "additionalProperties": true,
            "type": "object",
            "title": "Coverage",
            "description": "Coverage bitmap statistics"
          },
          "hindsight": {
            "additionalProperties": true,
            "type": "object",
            "title": "Hindsight",
            "description": "Hindsight feedback statistics"
          },
          "rate_limiter": {
            "additionalProperties": true,
            "type": "object",
            "title": "Rate Limiter",
            "description": "Adaptive rate limiter state"
          }
        },
        "type": "object",
        "required": [
          "state",
          "target",
          "elapsed",
          "requests_completed",
          "requests_failed",
          "requests_queued",
          "requests_dropped",
          "queue_size",
          "endpoints_found",
          "unique_urls",
          "transactions_stored",
          "phase_name",
          "current_phase",
          "coverage",
          "hindsight",
          "rate_limiter"
        ],
        "title": "CrawlStatusResponse",
        "description": "Full crawl status snapshot."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ModuleStatusResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "type": "string",
            "title": "State",
            "description": "Module state: pending, running, complete, error, needs_attention"
          },
          "requests_made": {
            "type": "integer",
            "title": "Requests Made",
            "default": 0
          },
          "endpoints_found": {
            "type": "integer",
            "title": "Endpoints Found",
            "default": 0
          },
          "errors": {
            "type": "integer",
            "title": "Errors",
            "default": 0
          },
          "duration_seconds": {
            "type": "number",
            "title": "Duration Seconds",
            "default": 0.0
          }
        },
        "type": "object",
        "required": [
          "name",
          "state"
        ],
        "title": "ModuleStatusResponse"
      },
...
      "PhaseStatusResponse": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "state": {
            "type": "string",
            "title": "State",
            "description": "Phase state: pending, running, complete, error, skipped"
          },
          "modules": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Modules"
          },
          "depends_on": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Depends On"
          },
          "parallel": {
            "type": "boolean",
            "title": "Parallel"
          },
          "exploration": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Exploration",
            "description": "Exploration stats: new_coverage, discovery_rate, requests, elapsed"
          }
        },
        "type": "object",
        "required": [
          "name",
          "state",
          "modules",
          "depends_on",
          "parallel"
        ],
        "title": "PhaseStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}